+++ src/string_func.h
@@ -45,10 +45,10 @@ static inline bool StrEmpty(const char *s) { return s == NULL || s[0] == '\0'; }
 
 
 /** Get the length of a string, within a limited buffer */
-static inline int ttd_strnlen(const char *str, int maxlen)
+static inline size_t ttd_strnlen(const char *str, size_t maxlen)
 {
 	const char *t;
-	for (t = str; *t != '\0' && t - str < maxlen; t++) {}
+	for (t = str; *t != '\0' && (size_t)(t - str) < maxlen; t++) {}
 	return t - str;
 }
